Refactor EditProfileForm and page export
