+++ src/app/api/auth/[...nextauth]/route.ts
@@ -54,7 +54,11 @@ export const authOptions: NextAuthOptions = {
         session.user.id = user.id;
         session.user.role = user.role;
         session.user.ChapterID = user.ChapterID;
+        session.user.firstName = user.firstName;
+        session.user.lastName = user.lastName;
+        session.user.pronouns = user.pronouns;
         session.user.name = user.firstName + " " + user.lastName;
+        session.user.email = user.email;
       }
       return session;
     },
