+++ src/components/user/index.tsx
@@ -1,1 +1,2 @@
 export { default as UserJoinRequest } from "./UserJoinRequest";
+export { default as EditProfileForm } from "./EditProfileForm";
